+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 11.1.0
+* Added rotateInPlace method to Sp3dObj, Sp3dFragment and Sp3dFace.
+* Added getCenter method to Sp3dFace.
+
 ## 11.0.0
 * Added isTouchable flag to Sp3dFragment.
 * Added setIsTouchableFlags method to Sp3dObj.